tondo_project/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
requests
gunicorn
faker
psycopg2>=2.9,<3.0
